+++ web/src/server.rs
@@ -19,10 +19,47 @@ enum Signal {
     Process,
 }
 
+#[derive(Copy, Clone, Debug)]
+pub enum SortOrder {
+    Up,
+    Down,
+}
+
+#[derive(Copy, Clone, Debug)]
+pub enum SortBy {
+    Pid,
+    Name,
+    Cpu,
+    Memory,
+}
+
+impl SortBy {
+    fn from(s: &str) -> SortBy {
+        match s {
+            "pid" => SortBy::Pid,
+            "name" => SortBy::Name,
+            "cpu" => SortBy::Cpu,
+            "mem" => SortBy::Memory,
+            _ => {
+                warn!("Unknown sort by: {}", s);
+                SortBy::Pid
+            }
+        }
+    }
+}
+
+#[derive(Copy, Clone, Debug)]
+pub struct Sort {
+    pub order: SortOrder,
+    pub by: SortBy
+}
+
 pub struct MyWebSocket {
     hb: Instant,
     sys: SystemInfo,
-    signal: Signal
+    signal: Signal,
+    pid: Option<String>,
+    sort: Option<Sort>
 }
 
 impl MyWebSocket {
@@ -30,7 +67,9 @@ impl MyWebSocket {
         Self {
             hb: Instant::now(),
             sys: SystemInfo::new(),
-            signal: Signal::Less
+            signal: Signal::Less,
+            pid: None,
+            sort: None
         }
     }
 
@@ -58,7 +97,7 @@ impl MyWebSocket {
                 match act.signal {
                     Signal::More => act.sys.get_full_fusion(),
                     Signal::Less => act.sys.get_less_fusion(),
-                    Signal::Process => act.sys.get_process_fusion(),
+                    Signal::Process => act.sys.get_process_fusion(act.pid.clone(), act.sort),
                 }
             )
         });
@@ -90,10 +129,35 @@ impl StreamHandler<Result<ws::Message, ws::ProtocolError>> for MyWebSocket {
                     let mut command = msg.splitn(2, ' ');
 
                     match command.next() {
-                        // let param = command.next();
                         Some("/more") => self.signal = Signal::More,
-                        Some("/less") => self.signal = Signal::Less,
-                        Some("/process") => self.signal = Signal::Process,
+                        Some("/less") => {
+                            self.signal = Signal::Less;
+                            // handle process
+                            self.pid = None;
+                            self.sort = None;
+                        },
+                        Some("/process") => {
+                            let param = command.next();
+                            self.signal = Signal::Process;
+                            self.pid = param.map(|s| s.to_string());
+                        },
+                        Some("/close_detail") => {
+                            self.pid = None;
+                        },
+                        Some("/up") => {
+                            let param = command.next();
+                            let sort = param.map(|sort_by| Sort {
+                                    order: SortOrder::Up,
+                                    by: SortBy::from(sort_by.to_lowercase().as_str())});
+                            self.sort = sort;
+                        },
+                        Some("/down") => {
+                            let param = command.next();
+                            let sort = param.map(|sort_by| Sort {
+                                order: SortOrder::Down,
+                                by: SortBy::from(sort_by.to_lowercase().as_str())});
+                            self.sort = sort;
+                        },
                         _ => {},
                     }
                 }